% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcg_data}
\alias{bcg_data}
\title{Example data for the background process}
\format{A dataset of 500 observations with 16 variables}
\usage{
data(bcg_data)
}
\description{
An example of the background data generated by mixture of 2 Gaussian components with equal proportions. Component parameters are specified
in detail in vignette (run ) The data is used to train a model to find anomalies in the experimental data
}
\references{
Working paper on a Statistical Learning method for Model-Independent searches for New Physics - AMVA4NewPhysics ITN
}
\keyword{datasets}
