% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_data}
\alias{experiment_data}
\title{Example data for the experimental data containing an unknown signal observations}
\format{A dataset of 500 observations with 16 variables}
\usage{
data(experiment_data)
}
\description{
An example of the experimental data generated by mixture background process and a signal process with
proportion equal to 10\%. The data could be used jointly with the background data to train a model and find anomalies in the data
at hand with respect to the background data.
}
\references{
Working paper on a Statistical Learning method for Model-Independent searches for New Physics - AMVA4NewPhysics ITN
}
\keyword{datasets}
