% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations2.R
\name{MAES}
\alias{MAES}
\title{MAES algorithm}
\usage{
MAES(data, J, labs = 0)
}
\arguments{
\item{data}{numeric matrix of size n and dimension P. The data on which the model is constructed.}

\item{J}{numeric(1). Number of Gaussian components used in the mixture to fit the model.}

\item{labs}{numeric vector of length equal to size of the data, default all labels are equal to 0. Used for computation of the classification results.}
}
\value{
resulting list with respective entries
\itemize{
\item means - matrix (JxP) of component means by row
\item eigenvectors - list of J component matrices containing eigenvectors by column
\item eigenvalues - list of J component diagonal matrices containing eigenvalues
\item proportions - component proportions \eqn{\pi}
\item gamma1 - \eqn{\gamma_1} mean regularization parameter
\item gamma2 - \eqn{\gamma_2} eigenvalue regularization parameter
\item ARI - adjusted Rand Index for classification results
\item pred_classes - model predicted classes
}
}
\description{
Method to obtain parameter estimates for the MAES algorithm. Slelection of regularization parameters is performed
automatically based on the modified BIC criteria
}
\examples{
data(bcg_data)

#Obtain model parameter estimates (might take about a minute)
f <- MAES(bcg_data, 2, attr(bcg_data, "label"))
}
\references{
Working paper on a Statistical Learning method for Model-Independent searches for New Physics - AMVA4NewPhysics ITN
}
