% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations2.R
\name{PAD}
\alias{PAD}
\title{Penalized anomaly detection}
\usage{
PAD(E_best, data_exp, gamma_eigen, gamma_mean, n_iter = 20, minbic, data)
}
\arguments{
\item{E_best}{MAES result. The background model - starting point for the algorithm, possible output model if signal is not found}

\item{data_exp}{numeric matrix, experimental data which is tested if contain anomaly}

\item{gamma_eigen}{numeric(1) regularization parameter \eqn{\gamma_2} for the component covariance eigenvalue penalty}

\item{gamma_mean}{numeric(1) regularization parameter \eqn{\gamma_1} for the mean parameters penalty}

\item{n_iter}{numeric(1) default equal to 20, maximum number of iteration for the algorithm outer loop}

\item{minbic}{numeric(1) the minimal BIC criteria reached so far.}

\item{data}{data.frame. background data against which anomalies are searched}

\item{minres}{numeric(1) I don't know}
}
\value{
resulting list with respective entries
\itemize{
\item means - matrix (JxP) of component means by row
\item eigenvectors - list of J component matrices containing eigenvectors by column
\item eigenvalues - list of J component diagonal matrices containing eigenvalues
\item proportions - component proportions \eqn{\pi}
\item gamma1 - \eqn{\gamma_1} mean regularization parameter
\item gamma2 - \eqn{\gamma_2} eigenvalue regularization parameter
}
}
\description{
Implementation of the PAD algorithm that enables to find the anomalous component in the experimental data.
}
\examples{
#Load the datasets
data(bcg_data)
data(experiment_data)

#1 fit the background model as a possible result if no anomaly is found
E_bcg=MAES(bcg_data,2,attr(bcg_data, "label"))

#2 fit the signal

#Iterate between fiting the bacground on data and fitting the signal on data2
minbic <- Inf
sig <- PAD(E_bcg, experiment_data,gamma_eigen = 1,
            gamma_mean = 4,minbic,bcg_data)

#Resulting component proportions
sig[[4]]
}
\references{
Working paper on a Statistical Learning method for Model-Independent searches for New Physics - AMVA4NewPhysics ITN
}
